PAL-618 Duplicated SimpleClient in Examples and Client repos: Removal of class that have been moved to Client and change needed to keep the example working.
