+++ example-model/mvn_dependency_tree.txt
@@ -100,14 +100,16 @@ uk.gov.gchq.palisade:example-model:jar:0.4.0-SNAPSHOT
 |     +- org.xerial.snappy:snappy-java:jar:1.1.1.3:compile
 |     +- org.apache.commons:commons-compress:jar:1.8.1:compile
 |     \- org.tukaani:xz:jar:1.5:compile
-+- uk.gov.gchq.palisade:example-library:jar:0.4.0-SNAPSHOT:compile
-|  +- uk.gov.gchq.palisade:hr-data-generator:jar:0.4.0-SNAPSHOT:compile
-|  |  +- com.github.javafaker:javafaker:jar:1.0.1:compile
-|  |  |  +- org.yaml:snakeyaml:jar:android:1.23:compile
-|  |  |  \- com.github.mifmif:generex:jar:1.0.2:compile
-|  |  |     \- dk.brics.automaton:automaton:jar:1.11-8:compile
-|  |  \- org.slf4j:slf4j-simple:jar:1.7.28:compile
++- uk.gov.gchq.palisade:clients-common:jar:0.4.0-SNAPSHOT:compile
++- uk.gov.gchq.palisade:clients-simple:jar:0.4.0-SNAPSHOT:compile
 |  \- uk.gov.gchq.palisade:readers-common:jar:0.4.0-SNAPSHOT:compile
++- uk.gov.gchq.palisade:example-library:jar:0.4.0-SNAPSHOT:compile
+|  \- uk.gov.gchq.palisade:hr-data-generator:jar:0.4.0-SNAPSHOT:compile
+|     +- com.github.javafaker:javafaker:jar:1.0.1:compile
+|     |  +- org.yaml:snakeyaml:jar:android:1.23:compile
+|     |  \- com.github.mifmif:generex:jar:1.0.2:compile
+|     |     \- dk.brics.automaton:automaton:jar:1.11-8:compile
+|     \- org.slf4j:slf4j-simple:jar:1.7.28:compile
 \- junit:junit:jar:4.12:test
    \- org.hamcrest:hamcrest-core:jar:2.1:test
       \- org.hamcrest:hamcrest:jar:2.1:test
